move db_error under db
