+++ src/agdb/collections/dictionary/dictionary_data.rs
@@ -1,6 +1,6 @@
 use super::dictionary_index::DictionaryIndex;
 use super::dictionary_value::DictionaryValue;
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::utilities::serialize::Serialize;
 use crate::utilities::stable_hash::StableHash;
