+++ src/agdb/collections/dictionary/dictionary_data_memory.rs
@@ -2,7 +2,7 @@ use super::dictionary_data::DictionaryData;
 use super::dictionary_index::DictionaryIndex;
 use super::dictionary_value::DictionaryValue;
 use crate::collections::multi_map::MultiMap;
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::utilities::serialize::Serialize;
 use crate::utilities::stable_hash::StableHash;
