+++ src/agdb/collections/dictionary/dictionary_data_storage.rs
@@ -1,6 +1,6 @@
 use crate::collections::storage_multi_map::StorageMultiMap;
 use crate::collections::storage_vec::StorageVec;
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::storage::storage_file::StorageFile;
 use crate::storage::storage_index::StorageIndex;
 use crate::storage::Storage;
