+++ src/agdb/collections/map_common/map_data_memory.rs
@@ -1,7 +1,7 @@
 use super::map_data::MapData;
 use super::map_value::MapValue;
 use super::map_value_state::MapValueState;
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::utilities::serialize::Serialize;
 use crate::utilities::stable_hash::StableHash;
 use std::hash::Hash;
