+++ src/agdb/collections/storage_vec.rs
@@ -1,5 +1,5 @@
 use super::vec::storage_vec_iterator::StorageVecIterator;
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::storage::storage_file::StorageFile;
 use crate::storage::storage_index::StorageIndex;
 use crate::storage::Storage;
