+++ src/agdb/graph/graph_data_storage.rs
@@ -2,7 +2,7 @@ use super::graph_data::GraphData;
 use super::graph_data_storage_indexes::GraphDataStorageIndexes;
 use super::graph_index::GraphIndex;
 use crate::collections::storage_vec::StorageVec;
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::storage::storage_file::StorageFile;
 use crate::storage::storage_index::StorageIndex;
 use crate::storage::Storage;
