+++ src/agdb/graph/storage_graph.rs
@@ -1,6 +1,6 @@
 use super::graph_data_storage::GraphDataStorage;
 use super::graph_impl::GraphImpl;
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::storage::storage_file::StorageFile;
 use crate::storage::storage_index::StorageIndex;
 use crate::storage::Storage;
