+++ src/agdb/storage.rs
@@ -8,7 +8,7 @@ mod storage_record;
 mod storage_records;
 mod write_ahead_log;
 
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::storage::storage_index::StorageIndex;
 use crate::utilities::serialize::Serialize;
