+++ src/agdb/storage/storage_index.rs
@@ -1,4 +1,4 @@
-use crate::db_error::DbError;
+use crate::db::db_error::DbError;
 use crate::utilities::serialize::Serialize;
 
 #[derive(Clone, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
